refactor: src folder
