+++ my_ciphering_cli/src/cipher.js
