+++ my_ciphering_cli/src/utils.js
@@ -1,15 +1,21 @@
-const {join} = require('path')
-const {access, constants} = require('fs')
+const { access, constants } = require('fs')
 
-const {NoSuchFileError, InvalidCipherPatternError} = require('./custom-errors')
-const {createCustomReadStream, createCustomWriteStream} = require('./streams/custom-streams')
 const cipher = require('./cipher')
+const { NoSuchFileError, InvalidCipherPatternError } = require('./custom-errors')
+const { createCustomReadStream, createCustomWriteStream } = require('../streams/custom-streams')
+
 const { 
     CaesarTransform, 
     Rot8Transform,
     AtbashTransform 
-} = require('./streams/transformer')
-const { CAESAR_SHIFT, ROT8_SHIFT, ATBASH_FLAG, MARKS } = require('./constants')
+} = require('../streams/transformer')
+
+const { 
+    CAESAR_SHIFT, 
+    ROT8_SHIFT, 
+    ATBASH_FLAG, 
+    MARKS 
+} = require('./constants')
 
 function errorHandler(err) {
     let { isCustom, name, message } = err
@@ -59,16 +65,31 @@ function sanitize(arg) {
 }
 
 async function generateReadStream(input) {
-    const isAccessable = await isFileAccessable(join(__dirname, input), 'r')
+    const isAccessable = await isFileAccessable(input, 'r')
     if (isAccessable) {
-        return createCustomReadStream(join(__dirname, input))
+        return createCustomReadStream(input)
     }
 }
 
 async function generateWriteStream(output) {
-    const isAccessable = await isFileAccessable(join(__dirname, output), 'w')
+    const isAccessable = await isFileAccessable(output, 'w')
     if (isAccessable)
-        return createCustomWriteStream(join(__dirname, output), { flags: 'a' })
+        return createCustomWriteStream(output, { flags: 'a' })
+}
+
+function consoleReadStream() {
+    process.stdout.write('Please enter your text here: \n')
+    process.stdin.resume()
+    return process.stdin
+}
+
+function pipelineErrorCb(err) {
+    if (err) {
+        process.stderr.write(err.message);
+        process.exit(1);
+    }
+    
+    console.log('Pipeline succeeded.');
 }
 
 module.exports = {
@@ -77,5 +98,7 @@ module.exports = {
     generateReadStream,
     generateWriteStream,
     generateStream,
-    sanitize
+    sanitize,
+    consoleReadStream,
+    pipelineErrorCb
 }
