+++ my_ciphering_cli/streams/custom-streams.test.js
@@ -1,14 +1,14 @@
 const fs = require('fs')
-const { createCustomReadStream } = require('./custom-streams')
+const { createCustomReadStream, createCustomWriteStream } = require('./custom-streams')
 
 describe('Custom Read and Write Streams', () => {
-    it('works', async () => {
+    it('custom read stream', () => {
         jest.spyOn(fs, 'open')
         jest.spyOn(fs, 'read')
-        const existingFilePath = './ss.txt'
+        const existingFilePath = './input.txt'
         
-        const rStream = await createCustomReadStream(existingFilePath)
-        rStream._construct() 
+        const rStream = createCustomReadStream(existingFilePath)
+        rStream._construct(() => {}) 
         rStream._read(13)
 
         expect(fs.open).toHaveBeenCalled()
@@ -16,5 +16,21 @@ describe('Custom Read and Write Streams', () => {
 
         fs.open.mockRestore()
         fs.read.mockRestore()
+    })
+
+    it('custom write stream', () => {
+        jest.spyOn(fs, 'open')
+        jest.spyOn(fs, 'write')
+        const existingFilePath = './output.txt'
+        
+        const wStream = createCustomWriteStream(existingFilePath, { flags: 'a' })
+        wStream._construct(() => {}) 
+        wStream._write('works', null, () => {})
+
+        expect(fs.open).toHaveBeenCalled()
+        expect(fs.write).toHaveBeenCalled()
+
+        fs.open.mockRestore()
+        fs.write.mockRestore()
     })
 })
